Make numatop.c free of global vars
numatop.c ultimately gets compiled into 'numatop' binary as it
contains the main() function. Thus, defining global variables
into this file will create issues if one wants to reuse other
files which uses these variables defined in numatop.c.

Move these variable definitions in appropriate files.
