+++ common/include/win.h
@@ -41,6 +41,11 @@
 extern "C" {
 #endif
 
+extern int g_sortkey;
+extern double g_llc_occupancy_scale;
+extern double g_llc_total_bw_scale;
+extern double g_llc_local_bw_scale;
+
 #define	NUMATOP_TITLE	"NumaTOP v2.0, (C) 2015 Intel Corporation"
 #define	CMD_CAPTION		"Command: "
 #define	WIN_PROCNAME_SIZE	12
